O4: Send imgs as uint8 to reduce upload payload by 53%
Cast imgs float32 → uint8 before msgpack serialization on the client;
server casts back to float32 before the camera encoder. Lossless since
CARLA captures uint8 pixels and _build_batch() only upcasts for local
inference.

Measured on loopback (n=168 ticks, vehicle agent):
  payload_kb:     8021 → 3803  (-4.2MB, -53%)
  to_numpy_ms:    0.04 → 0.31ms (+0.27ms — cheap uint8 cast)
  http_ms:        no change (upload not bottleneck on loopback)
  total_e2e_ms:   neutral (233ms → 233ms)

Kept despite no loopback benefit: on Azure over real network links,
4MB × 2 agents × 20Hz = 160 MB/s avoided upload bandwidth per
intersection. The loopback measurement is not representative of the
target deployment environment.

Also updates worldfusion_litserve_plan.md: O2/O3 marked tried+reverted
with data, O4 marked kept with rationale, full checklist audit.

+++ ecav/ml_manager/litserve_models.py
@@ -197,7 +197,8 @@ def _extract_wf_features(batch):
 
     if 'image_inputs' in batch and batch['image_inputs'] is not None:
         from einops import rearrange
-        imgs = batch['image_inputs']['imgs']
+        # Cast back to float32 if client sent as uint8 to reduce payload
+        imgs = batch['image_inputs']['imgs'].float()
         B, N, C, imH, imW = imgs.shape
         x = imgs.view(B * N, C, imH, imW)
         _, x = sensor.camenc(x, batch['image_inputs']['depth_map'], batch['record_len'])
